requirements.txt
django==4.2.0
psycopg2-binary==2.9.6
python-environ==0.4.54
django-cors-headers==3.14.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
djangorestframework==3.14.0
django-filter==23.1
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==7.0.2
pillow==9.5.0
django-phonenumbers==1.0.1
django-location-field
sorl-thumbnail==12.9.0
django-redis==5.2.0
Unidecode==1.3.6
django-image-uploader-widget
# eskiz-sms==0.2.3
django-redis-cache==0.13.1
celery
# django-ckeditor==6.5.1